+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/ComposerFormatSecurityContributor.java
@@ -17,6 +17,7 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 
 import org.sonatype.nexus.repository.Format;
+import org.sonatype.nexus.repository.composer.ComposerFormat;
 import org.sonatype.nexus.repository.security.RepositoryFormatSecurityContributor;
 
 /**
