+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/cleanup/ComposerCleanupPolicyConfiguration.java
@@ -18,7 +18,7 @@ import javax.inject.Named;
 import javax.inject.Singleton;
 
 import org.sonatype.nexus.cleanup.config.CleanupPolicyConfiguration;
-import org.sonatype.nexus.repository.composer.internal.ComposerFormat;
+import org.sonatype.nexus.repository.composer.ComposerFormat;
 
 import com.google.common.collect.ImmutableMap;
