+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/recipe/ComposerGroupRecipe.groovy
@@ -10,8 +10,13 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.composer.internal
+package org.sonatype.nexus.repository.composer.internal.recipe
 
+import org.sonatype.nexus.repository.composer.AssetKind
+import org.sonatype.nexus.repository.composer.ComposerFormat
+import org.sonatype.nexus.repository.composer.internal.ComposerGroupPackageJsonHandler
+import org.sonatype.nexus.repository.composer.internal.ComposerGroupPackagesJsonHandler
+import org.sonatype.nexus.repository.composer.internal.ComposerGroupProviderJsonHandler
 
 import javax.annotation.Nonnull
 import javax.inject.Inject
