+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/recipe/ComposerProxyRecipe.groovy
@@ -10,13 +10,13 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.composer.internal
+package org.sonatype.nexus.repository.composer.internal.recipe
 
+import org.sonatype.nexus.repository.composer.AssetKind
 import org.sonatype.nexus.repository.composer.ComposerFormat
-import org.sonatype.nexus.repository.composer.internal.ComposerHostedDownloadHandler
-import org.sonatype.nexus.repository.composer.ComposerHostedFacet
-import org.sonatype.nexus.repository.composer.internal.ComposerHostedUploadHandler
-import org.sonatype.nexus.repository.view.handlers.LastDownloadedHandler
+import org.sonatype.nexus.repository.composer.internal.ComposerPackageHandler
+import org.sonatype.nexus.repository.composer.internal.ComposerProviderHandler
+import org.sonatype.nexus.repository.composer.internal.ComposerProxyFacet
 
 import javax.annotation.Nonnull
 import javax.inject.Inject
@@ -27,41 +27,53 @@ import javax.inject.Singleton
 import org.sonatype.nexus.repository.Format
 import org.sonatype.nexus.repository.Repository
 import org.sonatype.nexus.repository.Type
+import org.sonatype.nexus.repository.cache.NegativeCacheFacet
+import org.sonatype.nexus.repository.cache.NegativeCacheHandler
 import org.sonatype.nexus.repository.http.HttpHandlers
-import org.sonatype.nexus.repository.types.HostedType
+import org.sonatype.nexus.repository.httpclient.HttpClientFacet
+import org.sonatype.nexus.repository.proxy.ProxyHandler
+import org.sonatype.nexus.repository.purge.PurgeUnusedFacet
+import org.sonatype.nexus.repository.types.ProxyType
 import org.sonatype.nexus.repository.view.ConfigurableViewFacet
 import org.sonatype.nexus.repository.view.Router
 import org.sonatype.nexus.repository.view.ViewFacet
 
-import static org.sonatype.nexus.repository.composer.internal.AssetKind.PACKAGE
-import static org.sonatype.nexus.repository.composer.internal.AssetKind.PACKAGES
-import static org.sonatype.nexus.repository.composer.internal.AssetKind.PROVIDER
-import static org.sonatype.nexus.repository.composer.internal.AssetKind.ZIPBALL
-
 /**
- * Recipe for creating a Composer hosted repository.
+ * Recipe for creating a Composer proxy repository.
  */
-@Named(ComposerHostedRecipe.NAME)
+@Named(ComposerProxyRecipe.NAME)
 @Singleton
-class ComposerHostedRecipe
+class ComposerProxyRecipe
     extends ComposerRecipeSupport
 {
-  public static final String NAME = 'composer-hosted'
+  public static final String NAME = 'composer-proxy'
 
   @Inject
-  Provider<ComposerHostedFacet> hostedFacet
+  Provider<ComposerProxyFacet> proxyFacet
 
   @Inject
-  LastDownloadedHandler lastDownloadedHandler
+  Provider<NegativeCacheFacet> negativeCacheFacet
 
   @Inject
-  ComposerHostedUploadHandler uploadHandler
+  Provider<PurgeUnusedFacet> purgeUnusedFacet
 
   @Inject
-  ComposerHostedDownloadHandler downloadHandler
+  NegativeCacheHandler negativeCacheHandler
 
   @Inject
-  ComposerHostedRecipe(@Named(HostedType.NAME) final Type type, @Named(ComposerFormat.NAME) final Format format) {
+  ProxyHandler proxyHandler
+
+  @Inject
+  Provider<HttpClientFacet> httpClientFacet
+
+  @Inject
+  ComposerProviderHandler composerProviderHandler
+
+  @Inject
+  ComposerPackageHandler composerPackageHandler
+
+  @Inject
+  ComposerProxyRecipe(@Named(ProxyType.NAME) final Type type, @Named(ComposerFormat.NAME) final Format format) {
     super(type, format)
   }
 
@@ -70,79 +82,83 @@ class ComposerHostedRecipe
     repository.attach(contentFacet.get())
     repository.attach(securityFacet.get())
     repository.attach(configure(viewFacet.get()))
-    repository.attach(hostedFacet.get())
-    repository.attach(maintenanceFacet.get())
+    repository.attach(httpClientFacet.get())
+    repository.attach(negativeCacheFacet.get())
+    repository.attach(proxyFacet.get())
     repository.attach(searchFacet.get())
     repository.attach(browseFacet.get())
+    repository.attach(purgeUnusedFacet.get())
+    repository.attach(maintenanceFacet.get())
   }
 
-  /**
-   * Configure {@link ViewFacet}.
-   */
   private ViewFacet configure(final ConfigurableViewFacet facet) {
     Router.Builder builder = new Router.Builder()
 
     builder.route(packagesMatcher()
         .handler(timingHandler)
-        .handler(assetKindHandler.rcurry(PACKAGES))
+        .handler(assetKindHandler.rcurry(AssetKind.PACKAGES))
         .handler(securityHandler)
         .handler(exceptionHandler)
         .handler(handlerContributor)
+        .handler(negativeCacheHandler)
         .handler(conditionalRequestHandler)
         .handler(partialFetchHandler)
         .handler(contentHeadersHandler)
-        .handler(downloadHandler)
+        .handler(proxyHandler)
+        .create())
+
+    builder.route(listMatcher()
+        .handler(timingHandler)
+        .handler(assetKindHandler.rcurry(AssetKind.LIST))
+        .handler(securityHandler)
+        .handler(exceptionHandler)
+        .handler(handlerContributor)
+        .handler(negativeCacheHandler)
+        .handler(conditionalRequestHandler)
+        .handler(partialFetchHandler)
+        .handler(contentHeadersHandler)
+        .handler(proxyHandler)
         .create())
 
     builder.route(providerMatcher()
         .handler(timingHandler)
-        .handler(assetKindHandler.rcurry(PROVIDER))
+        .handler(assetKindHandler.rcurry(AssetKind.PROVIDER))
         .handler(securityHandler)
         .handler(exceptionHandler)
         .handler(handlerContributor)
+        .handler(negativeCacheHandler)
         .handler(conditionalRequestHandler)
         .handler(partialFetchHandler)
         .handler(contentHeadersHandler)
-        .handler(lastDownloadedHandler)
-        .handler(downloadHandler)
+        .handler(composerProviderHandler)
+        .handler(proxyHandler)
         .create())
 
     builder.route(packageMatcher()
-        .handler(timingHandler)
-        .handler(assetKindHandler.rcurry(PACKAGE))
-        .handler(securityHandler)
-        .handler(exceptionHandler)
-        .handler(handlerContributor)
-        .handler(conditionalRequestHandler)
-        .handler(partialFetchHandler)
-        .handler(contentHeadersHandler)
-        .handler(lastDownloadedHandler)
-        .handler(downloadHandler)
-        .create())
+            .handler(timingHandler)
+            .handler(assetKindHandler.rcurry(AssetKind.PACKAGE))
+            .handler(securityHandler)
+            .handler(exceptionHandler)
+            .handler(handlerContributor)
+            .handler(negativeCacheHandler)
+            .handler(conditionalRequestHandler)
+            .handler(partialFetchHandler)
+            .handler(contentHeadersHandler)
+            .handler(composerPackageHandler)
+            .handler(proxyHandler)
+            .create())
 
     builder.route(zipballMatcher()
         .handler(timingHandler)
-        .handler(assetKindHandler.rcurry(ZIPBALL))
+        .handler(assetKindHandler.rcurry(AssetKind.ZIPBALL))
         .handler(securityHandler)
         .handler(exceptionHandler)
         .handler(handlerContributor)
+        .handler(negativeCacheHandler)
         .handler(conditionalRequestHandler)
         .handler(partialFetchHandler)
         .handler(contentHeadersHandler)
-        .handler(lastDownloadedHandler)
-        .handler(downloadHandler)
-        .create())
-
-    builder.route(uploadMatcher()
-        .handler(timingHandler)
-        .handler(assetKindHandler.rcurry(ZIPBALL))
-        .handler(securityHandler)
-        .handler(exceptionHandler)
-        .handler(handlerContributor)
-        .handler(conditionalRequestHandler)
-        .handler(partialFetchHandler)
-        .handler(contentHeadersHandler)
-        .handler(uploadHandler)
+        .handler(proxyHandler)
         .create())
 
     addBrowseUnsupportedRoute(builder)
