+++ nexus-repository-composer/src/main/java/org/sonatype/nexus/repository/composer/internal/recipe/ComposerRecipeSupport.groovy
@@ -10,10 +10,10 @@
  * of Sonatype, Inc. Apache Maven is a trademark of the Apache Software Foundation. M2eclipse is a trademark of the
  * Eclipse Foundation. All other trademarks are the property of their respective owners.
  */
-package org.sonatype.nexus.repository.composer.internal
+package org.sonatype.nexus.repository.composer.internal.recipe
 
-import org.sonatype.nexus.repository.composer.internal.AssetKind
-import org.sonatype.nexus.repository.composer.internal.ComposerContentFacet
+import org.sonatype.nexus.repository.composer.AssetKind
+import org.sonatype.nexus.repository.composer.ComposerContentFacet
 import org.sonatype.nexus.repository.composer.internal.ComposerMaintenanceFacet
 import org.sonatype.nexus.repository.composer.internal.ComposerSecurityFacet
 import org.sonatype.nexus.repository.content.browse.BrowseFacet
